test: QuackButton 스냅샷 테스트 추가
